+++ CHANGELOG.md
@@ -3,6 +3,12 @@ project changelog
 
 All notable changes will be documented in this file
 
+[1.1.0] - 08.03.2019
+Added
+* react example
+Changed
+* remove ES5 transpilation and webpack bundling. now will be distributed ES6
+
 [1.0.0] - 23.12.2018
 
 * yo!